Improve desloppify code health
